[mod/#31] 평가 데이터 초기화 데이터 수정 및 UiState, 모델 분리
